Moving tests around
