+++ test/integration/websocket_integration_test.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "test/integration/http_integration.h"
+
+#include "gtest/gtest.h"
+
+namespace Envoy {
+class WebsocketIntegrationTest : public HttpIntegrationTest,
+                                 public testing::TestWithParam<Network::Address::IpVersion> {
+public:
+  WebsocketIntegrationTest() : HttpIntegrationTest(Http::CodecClient::Type::HTTP1, GetParam()) {}
+};
+} // namespace Envoy
